Combine RT Tests Under Single Submodule
