+++ tests/tests/ray_tracing/scene/mesh_gen.rs
