+++ tests/tests/root.rs
@@ -34,7 +34,6 @@ mod push_constants;
 mod query_set;
 mod queue_transfer;
 mod ray_tracing;
-mod ray_tracing_use_after_free;
 mod render_pass_ownership;
 mod resource_descriptor_accessor;
 mod resource_error;
